chore: Extract commands into separate modules
